fp16<->bf16

+++ fms/modules/attention.py
@@ -498,11 +498,11 @@ class MultiHeadAttention(nn.Module):
         queries = queries.unflatten(2, (self.kvheads, expansion))  # b l h e d
 
         # b l h e d, b n h d, l c
-        attn = self.indlinear(queries, keys, self.plan[-1])  # b l h e c
+        attn = self.indlinear(queries.to(dtype=torch.float16), keys.to(dtype=torch.float16), self.plan[-1])  # b l h e c
         attn = attn.softmax(4)
-        attn = self.indlineart(attn, values, self.plan[-1])  # b l h e d
+        attn = self.indlineart(attn, values.to(dtype=torch.float16), self.plan[-1])  # b l h e d
 
-        attn = attn.reshape(batch_size, q_len, self.nheads * self.emb_v_per_head)
+        attn = attn.to(dtype=torch.bfloat16).reshape(batch_size, q_len, self.nheads * self.emb_v_per_head)
         out = self.dense(attn)
 
         # if use_cache=True, we return the hidden_state as well as the kv cache
